+++ src/main/java/com/example/recruitment_service/repository/JobProvinceRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.math.BigInteger;
 
 @Repository
-public interface JobProvinceRepository extends JpaRepository<JobProvince, BigInteger> {
+public interface JobProvinceRepository extends JpaRepository<JobProvince, Integer> {
 
 }
